requirements.txt
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
flask
flask-cors
nltk